backend/requirements.txt
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
SQLAlchemy==2.0.23